QOL refactor - renamed sidebar components so they're next to each other in the filesystem.  did not bother to rename the subcomponents, sue me.
